pyproject.toml
[tool.poetry]
name = "gs_tools_django"
version = "0.1.0"
description = ""
authors = ["mahdi13830510 <mahdi.dlt13830510@Gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg = [
    { platform = "win32", extras = ["binary"], version = "^3.1.10" },
    { platform = "linux", extras = ["c"], version = "^3.1.10" }
]
django = {extras = ["argon2"], version = "^5.1.7"}
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
gunicorn = "^23.0.0"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^8.0.0" }
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
django-health-check = "^3.18.3"
django-cacheops = "^7.1"
celery = "^5.4.0"
redis = "^5.2.0"
orjson = "^3.10.11"
django-storages = "^1.14.4"
coverage = "^7.6.8"
python-magic = { platform = "linux", version = "^0.4.27" }
python-magic-bin = { platform = "win32", version = "^0.4.14" }
factory-boy = "^3.3.1"
structlog = "^24.4.0"
django-structlog = { extras = ["celery"], version = "^9.0.0" }
djangorestframework-simplejwt = "^5.3.1"
python-calamine = "^0.3.1"
httpx = "^0.28.1"
django-extensions = "^3.2.3"
django-debug-toolbar = "^5.1.0"
ujson = "^5.10.0"
ptpython = "^3.0.29"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"
pre-commit = "^4.0.1"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
